EE-296: inital data model change and migration (#213)
* EE-296: inital data model change and migration

* EE-296: updated project agrigation function

* EE-296
