+++ api/controllers/search.js
@@ -172,7 +172,7 @@ var handleDateEndItem = function (expArray, field, entry) {
   }
 }
 
-var searchCollection = async function (roles, keywords, collection, pageNum, pageSize, project, sortField = undefined, sortDirection = undefined, caseSensitive, populate = false, and, or) {
+var searchCollection = async function (roles, keywords, collection, pageNum, pageSize, project, projectLegislation = undefined, sortField = undefined, sortDirection = undefined, caseSensitive, populate = false, and, or) {
   var properties = undefined;
   if (project) {
     properties = { project: mongoose.Types.ObjectId(project) };
@@ -280,21 +280,87 @@ var searchCollection = async function (roles, keywords, collection, pageNum, pag
   }
 
   if (collection === 'Project') {
-    // pop proponent if exists.
-    aggregation.push(
-      {
-        '$lookup': {
-          "from": "epic",
-          "localField": "proponent",
-          "foreignField": "_id",
-          "as": "proponent"
-        }
+    var projectDataId;
+
+    switch (projectLegislation) {
+      case "1996", "2002", "2018":
+        projectDataId = "allProjectData[" + projectLegislation + "]";
+        break;
+      case "all":
+        projectDataId = [ "allProjectData[1996]", "allProjectData[2002]", "allProjectData[2018]" ];
+        break;
+      default:
+        projectDataId = "currentProjectData";
+        break;
+    }
+
+    if (projectLegislation === "all"){
+      projectDataId.foreach ( dataId => {
+        // pop most recent project data.
+        aggregation.push(
+          {
+            '$lookup': {
+              "from": "epic",
+              "localField": dataId,
+              "foreignField": "_id",
+              "as": dataId
+            }
+          });
+        aggregation.push(
+          {
+            "$unwind": "$" + dataId
+          },
+        );
+
+        // pop proponent if exists.
+        aggregation.push(
+          {
+            '$lookup': {
+              "from": "epic",
+              "localField": dataId + ".proponent",
+              "foreignField": "_id",
+              "as": dataId + ".proponent"
+            }
+          });
+        aggregation.push(
+          {
+            "$unwind": "$" + dataId + ".proponent"
+          },
+        );
       });
-    aggregation.push(
-      {
-        "$unwind": "$proponent"
-      },
-    );
+    } else {
+      // pop most recent project data.
+      aggregation.push(
+        {
+          '$lookup': {
+            "from": "epic",
+            "localField": projectDataId,
+            "foreignField": "_id",
+            "as": projectDataId
+          }
+        });
+      aggregation.push(
+        {
+          "$unwind": "$" + projectDataId
+        },
+      );
+
+      // pop proponent if exists.
+      aggregation.push(
+        {
+          '$lookup': {
+            "from": "epic",
+            "localField": projectDataId + ".proponent",
+            "foreignField": "_id",
+            "as": projectDataId + ".proponent"
+          }
+        });
+      aggregation.push(
+        {
+          "$unwind": "$" + projectDataId + ".proponent"
+        },
+      );
+    }
   }
 
   if (collection === 'Group') {
